requirements.txt
Flask==3.0.3
gunicorn==22.0.0
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
